requirements.txt
flask
pillow
APScheduler
Waitress
PyYaml